# docs/api-reference/openapi.yaml
openapi: 3.0.3
info:
  title: Retail POS SaaS API
  version: 1.0.0
  description: Multi-tenant backend for fashion retail point of sale.
paths:
...
  /api/v1/sales/:
    get:
      operationId: sales_list
      description: |-
        Sales are created and cancelled, never edited.

        A completed sale is an accounting fact. Correcting it means a refund or a
        cancellation, both of which leave a trail.
      parameters:
      - in: query
        name: cash_session
        schema:
          type: string
          format: uuid
      - in: query
        name: customer
        schema:
          type: string
          format: uuid
      - in: query
        name: location
        schema:
          type: string
          format: uuid
      - in: query
        name: occurred_after
        schema:
          type: string
          format: date-time
      - in: query
        name: occurred_before
        schema:
          type: string
          format: date-time
      - name: ordering
        required: false
        in: query
        description: Qué campo usar para ordenar los resultados.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Número de resultados a devolver por página.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: Un término de búsqueda.
        schema:
          type: string
      - in: query
        name: seller
        schema:
          type: string
          format: uuid
      - in: query
        name: source
        schema:
          type: string
          enum:
          - POS
          - SYNC
        description: |-
          * `POS` - Point of sale
          * `SYNC` - Synced from an offline terminal
      - in: query
        name: status
        schema:
          type: string
          enum:
          - CANCELLED
          - COMPLETED
          - DRAFT
          - PARTIALLY_REFUNDED
          - REFUNDED
        description: |-
          * `DRAFT` - Draft
          * `COMPLETED` - Completed
          * `CANCELLED` - Cancelled
          * `REFUNDED` - Refunded
          * `PARTIALLY_REFUNDED` - Partially refunded
      tags:
      - sales
      security:
      - organizationJWT: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSaleListList'
          description: ''
    post:
      operationId: sales_create
      description: |-
        Sales are created and cancelled, never edited.

        A completed sale is an accounting fact. Correcting it means a refund or a
        cancellation, both of which leave a trail.
      tags:
      - sales
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SaleCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SaleCreateRequest'
        required: true
      security:
      - organizationJWT: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
          description: ''
...
components:
  schemas:
...
    PaginatedSaleListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/SaleList'
...
    Payment:
      type: object
      description: |-
        Base serializer for tenant-owned models.

        Overriding `serializer_related_field` makes every auto-generated relation
        tenant-scoped, so isolation does not depend on each developer remembering
        to scope a nested id by hand. `organization` is never writable: it comes
        from the authenticated context.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        method:
          allOf:
          - $ref: '#/components/schemas/PaymentMethodEnum'
          readOnly: true
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        reference:
          type: string
          readOnly: true
          description: Voucher, approval code.
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - amount
      - created_at
      - id
      - method
      - reference
    PaymentInputRequest:
      type: object
      properties:
        method:
          $ref: '#/components/schemas/PaymentMethodEnum'
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
        reference:
          type: string
          maxLength: 80
      required:
      - amount
      - method
    PaymentMethodEnum:
      enum:
      - CASH
      - CARD
      - TRANSFER
      - OTHER
      type: string
      description: |-
        * `CASH` - Efectivo
        * `CARD` - Tarjeta
        * `TRANSFER` - Transferencia
        * `OTHER` - Otro
...
    Refund:
      type: object
      description: |-
        Base serializer for tenant-owned models.

        Overriding `serializer_related_field` makes every auto-generated relation
        tenant-scoped, so isolation does not depend on each developer remembering
        to scope a nested id by hand. `organization` is never writable: it comes
        from the authenticated context.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        number:
          type: string
          readOnly: true
        sale:
          type: string
          format: uuid
          readOnly: true
        sale_number:
          type: string
          readOnly: true
        location:
          type: string
          format: uuid
          readOnly: true
        cash_session:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        method:
          allOf:
          - $ref: '#/components/schemas/PaymentMethodEnum'
          readOnly: true
        restock:
          type: boolean
          readOnly: true
          description: False when the goods come back damaged and cannot be resold.
        reason:
          type: string
          readOnly: true
        occurred_at:
          type: string
          format: date-time
          readOnly: true
        created_by:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/RefundItem'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - cash_session
      - created_at
      - created_by
      - id
      - items
      - location
      - method
      - number
      - occurred_at
      - reason
      - restock
      - sale
      - sale_number
      - total
...
    RefundItem:
      type: object
      description: |-
        Base serializer for tenant-owned models.

        Overriding `serializer_related_field` makes every auto-generated relation
        tenant-scoped, so isolation does not depend on each developer remembering
        to scope a nested id by hand. `organization` is never writable: it comes
        from the authenticated context.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        sale_item:
          type: string
          format: uuid
          readOnly: true
        description:
          type: string
          readOnly: true
        quantity:
          type: integer
          readOnly: true
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
      required:
      - amount
      - description
      - id
      - quantity
      - sale_item
...
    Sale:
      type: object
      description: |-
        Base serializer for tenant-owned models.

        Overriding `serializer_related_field` makes every auto-generated relation
        tenant-scoped, so isolation does not depend on each developer remembering
        to scope a nested id by hand. `organization` is never writable: it comes
        from the authenticated context.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        number:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/SaleStatusEnum'
          readOnly: true
        location:
          type: string
          format: uuid
          readOnly: true
        customer:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        customer_name:
          type: string
          readOnly: true
        seller:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        cash_session:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        discount_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        tax_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        paid_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        change_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        refunded_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        net_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        source:
          allOf:
          - $ref: '#/components/schemas/SourceEnum'
          readOnly: true
        device_id:
          type: string
          readOnly: true
          description: Terminal that created it offline.
        occurred_at:
          type: string
          format: date-time
          readOnly: true
        notes:
          type: string
          readOnly: true
        cancelled_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        cancellation_reason:
          type: string
          readOnly: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleItem'
          readOnly: true
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
          readOnly: true
        refunds:
          type: array
          items:
            $ref: '#/components/schemas/Refund'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - cancellation_reason
      - cancelled_at
      - cash_session
      - change_amount
      - created_at
      - customer
      - customer_name
      - device_id
      - discount_total
      - id
      - items
      - location
      - net_total
      - notes
      - number
      - occurred_at
      - paid_total
      - payments
      - refunded_total
      - refunds
      - seller
      - source
      - status
      - subtotal
      - tax_total
      - total
...
    SaleCreateRequest:
      type: object
      description: A complete sale in one call. Totals are always recomputed server-side.
      properties:
        id:
          type: string
          format: uuid
          description: Optional client-generated id, so an offline terminal keeps
            the id it printed.
        location:
          type: string
          format: uuid
        customer:
          type: string
          format: uuid
          nullable: true
        cash_register:
          type: string
          format: uuid
          nullable: true
        occurred_at:
          type: string
          format: date-time
        notes:
          type: string
        expected_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          nullable: true
          description: If sent and it disagrees with the server total, the sale is
            rejected (409).
        lines:
          type: array
          items:
            $ref: '#/components/schemas/SaleLineInputRequest'
        payments:
          type: array
          items:
            $ref: '#/components/schemas/PaymentInputRequest'
      required:
      - lines
      - payments
    SaleItem:
      type: object
      description: |-
        Base serializer for tenant-owned models.

        Overriding `serializer_related_field` makes every auto-generated relation
        tenant-scoped, so isolation does not depend on each developer remembering
        to scope a nested id by hand. `organization` is never writable: it comes
        from the authenticated context.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        variant:
          type: string
          format: uuid
          readOnly: true
        sku:
          type: string
          readOnly: true
        description:
          type: string
          readOnly: true
        quantity:
          type: integer
          readOnly: true
        unit_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        discount_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        line_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        tax_rate:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          readOnly: true
        taxable_base:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        tax_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        refunded_quantity:
          type: integer
          readOnly: true
        refundable_quantity:
          type: integer
          readOnly: true
      required:
      - description
      - discount_amount
      - id
      - line_total
      - quantity
      - refundable_quantity
      - refunded_quantity
      - sku
      - tax_amount
      - tax_rate
      - taxable_base
      - unit_price
      - variant
    SaleLineInputRequest:
      type: object
      properties:
        variant:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        unit_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          nullable: true
          description: Overrides the catalogue price. Omit to use the shelf price.
        discount_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          default: '0.00'
      required:
      - quantity
      - variant
    SaleList:
      type: object
      description: |-
        Base serializer for tenant-owned models.

        Overriding `serializer_related_field` makes every auto-generated relation
        tenant-scoped, so isolation does not depend on each developer remembering
        to scope a nested id by hand. `organization` is never writable: it comes
        from the authenticated context.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        number:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/SaleStatusEnum'
          readOnly: true
        location:
          type: string
          format: uuid
          readOnly: true
        customer:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        customer_name:
          type: string
          readOnly: true
        seller:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        seller_email:
          type: string
          readOnly: true
        total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        refunded_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        item_count:
          type: integer
          readOnly: true
        payment_methods:
          type: array
          items:
            type: string
          readOnly: true
        occurred_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - customer
      - customer_name
      - id
      - item_count
      - location
      - number
      - occurred_at
      - payment_methods
      - refunded_total
      - seller
      - seller_email
      - status
      - total
    SaleStatusEnum:
      enum:
      - DRAFT
      - COMPLETED
      - CANCELLED
      - REFUNDED
      - PARTIALLY_REFUNDED
      type: string
      description: |-
        * `DRAFT` - Draft
        * `COMPLETED` - Completed
        * `CANCELLED` - Cancelled
        * `REFUNDED` - Refunded
        * `PARTIALLY_REFUNDED` - Partially refunded
...
    SourceEnum:
      enum:
      - POS
      - SYNC
      type: string
      description: |-
        * `POS` - Point of sale
        * `SYNC` - Synced from an offline terminal
...
  securitySchemes:
    organizationJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from /api/v1/auth/login/. An account belongs to exactly
        one business, so the tenant is derived from the account on every request -
        there is nothing to select and nothing to switch.
servers:
- url: http://localhost:8000
  description: Local development